+++ R/map.R
@@ -29,9 +29,9 @@
 #'   objects, a string indicating the name of a logical element in the
 #'   inner lists. Only those elements where `.p` evaluates to
 #'   `TRUE` will be modified.
-#' @param .at A character vector of names or a numeric vector of
-#'   positions. Only those elements corresponding to `.at` will be
-#'   modified.
+#' @param .at A character vector of names, positive numeric vector of
+#'   positions to include, or a negative numeric vector of positions to
+#'   exlude. Only those elements corresponding to `.at` will be modified.
 #' @param ... Additional arguments passed on to `.f`.
 #' @return All functions return a vector the same length as `.x`.
 #'