+++ tests/testthat/test-map.R
@@ -72,6 +72,12 @@ test_that("map_if() and map_at() always return a list", {
   expect_identical(map_at(df, 1, ~"out"), list(x = "out", y = "a"))
 })
 
+test_that("negative .at omits locations", {
+  x <- c(1, 2, 3)
+  out <- map_at(x, -1, ~ .x * 2)
+  expect_equal(out, list(1, 4, 6))
+})
+
 test_that("map works with calls and pairlists", {
   out <- map(quote(f(x)), ~ quote(z))
   expect_equal(out, list(quote(z), quote(z)))